bubblewrap: Support SKIP_CLEANUP
Useful for debugging because it allows to just copy and paste the bwrap
command from the test log to get a shell in the bubblewrap sandbox.

+++ internal/users/locking/locking_bwrap_test.go
@@ -479,7 +479,9 @@ func TestMain(m *testing.M) {
 	if err != nil {
 		panic(err)
 	}
-	defer func() { _ = os.RemoveAll(tempDir) }()
+	if v := os.Getenv("SKIP_CLEANUP"); v == "" {
+		defer func() { _ = os.RemoveAll(tempDir) }()
+	}
 
 	m.Run()
 }
